# config/swagger.json
{
  "swagger": "2.0",
  "info": {
    "description": "This is ruckus AP registration system to bring up new APs",
    "version": "1.0.0",
    "title": "Ruckus Accesspoint Registration",
    "termsOfService": "http://ruckuswireless.com/terms/",
    "contact": {
      "name": "Ruckus",
      "email": "support@ruckuswireless.com"
    },
    "license": {
      "name": "License Ruckus wireless system",
      "url": "http://www.apache.org/licenses/LICENSE-2.0.html"
    }
  },
  "basePath": "/",
...
  "schemes": ["https"],
  "paths": {
...
	"/api/adminactivitieslist": {
      "get": {
        "tags": ["Admin Activity Logs"],
        "summary": "Get List of admin activities",
        "description": "",
        "operationId": "adminactivitieslist",
        "consumes": ["application/json"],
        "produces": ["application/json"],
        "parameters": [{
          "name": "startindex",
          "in": "query",
          "description": "start index",
          "required": false,
          "type": "integer",
		  "default": 1
		},
		{
          "name": "numberofrows",
          "in": "query",
          "description": "number of rows",
          "required": false,
          "type": "integer",
		  "default": 10
		},
		{
          "name": "sortby",
          "in": "query",
          "description": "Sort By Column",
          "required": false,
          "type": "string",
		  "default": "activitytime"
		},
		{
          "name": "sortorder",
          "in": "query",
          "description": "Sort By Order [1 => Asc, -1 => Desc]",
          "required": false,
          "type": "integer",
		  "default": -1
		},
		{
          "name": "search",
          "in": "query",
          "description": "Search text",
          "required": false,
          "type": "string"
		}
		],
        "responses": {
          "405": {
            "description": "Invalid input"
          }
        },
        "security":[{
          "xAaccessToken": []
        }]
      }
    }
  },
  "securityDefinitions": {
    "xAaccessToken": {
      "type": "apiKey",
      "name": "x-access-token",
      "in": "header"
    }
  },
...
}